aula6 - session management/app/requirements.txt
Flask==2.3.3
PyJWT==1.7.1
Jinja2==3.1.2
Werkzeug==2.3.7
